+++ CHANGELOG.md
@@ -1,2 +1,6 @@
+# 2.0.10
+- Read `biomesMerging.fromCavesGeneration.enabled` correctly from config.
+- Fix cave biome that where over surface on eroded biome.
+
 # 2.0.9
 - Config: Aera values will be swap if min > max.